+++ src/services/form-validation.ts
@@ -240,8 +240,7 @@ type ValidateJSSchema = Record<string, unknown>
 const presence = (
   { required, requiredMessage }: FormTypes.FormElementRequired,
   message: string,
-) =>
-  required ? { message: requiredMessage || message, allowEmpty: false } : false
+) => (required ? { message: requiredMessage || message } : false)
 
 const escapeElementName = (elementName: string) => {
   const escapedName = elementName.replace(/\./g, '\\.')