+++ src/form-elements/FormElementSignature.tsx
@@ -13,7 +13,7 @@ import AttachmentStatus from '../components/renderer/attachments/AttachmentStatu
 import { canvasToBlob } from '../services/blob-utils'
 import ImagePreviewUnavailable from '../components/renderer/attachments/ImagePreviewUnavailable'
 import { FormElementValueChangeHandler } from '../types/form'
-import useIsPageVisible from '../hooks/usePage'
+import usePage from '../hooks/usePage'
 import ProgressBar from '../components/renderer/attachments/ProgressBar'
 import { IsDirtyProps } from '../types/form'
 
@@ -36,7 +36,7 @@ function FormElementSignature({
   setIsDirty,
   isDirty,
 }: Props) {
-  const isPageVisible = useIsPageVisible()
+  const { isPageVisible } = usePage()
 
   const handleChange = React.useCallback<
     FormElementValueChangeHandler<FormElementBinaryStorageValue>