+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- onBlur validation for `boolean`, `camera`, `checkboxes`, `files`, `location` and `radio` elements
+
 ## [6.2.2] - 2024-07-10
 
 ## [6.2.1] - 2024-07-01