+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [6.2.2] - 2024-07-10
 
+### Added
+
+- bottom collapse button to sections
+
 ## [6.2.1] - 2024-07-01
 
 ### Changed