+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - onBlur validation for `boolean`, `camera`, `checkboxes`, `files`, `location` and `radio` elements
 - elements with options allowing empty strings
+- threeDS2 in quickstream trusted frame
 
 ## [6.2.2] - 2024-07-10